Refactor IDs to long and add EF repository implementations

+++ GardenShop.Api.Application/Abstractions/IRepositories/IOrderRepository.cs
@@ -5,7 +5,7 @@ namespace GardenShop.Application.Abstractions.Persistence
 	public interface IOrderRepository : IRepository<Order>
 	{
 		Task<List<Order>> GetAllAsync(CancellationToken ct);
-		Task<Order?> GetByIdWithItemsAsync(int id, CancellationToken ct);
-		Task<Order?> GetByIdWithItemsAndProductsAsync(int id, CancellationToken ct);
+		Task<Order?> GetByIdWithItemsAsync(long id, CancellationToken ct);
+		Task<Order?> GetByIdWithItemsAndProductsAsync(long id, CancellationToken ct);
 	}
 }
