+++ GardenShop.Api.Application/Abstractions/IRepositories/IRepository.cs
@@ -4,7 +4,7 @@ namespace GardenShop.Application.Abstractions.Persistence
 {
 	public interface IRepository<T> where T : BaseEntity
 	{
-		Task<T?> GetByIdAsync(int id, CancellationToken ct);
+		Task<T?> GetByIdAsync(long id, CancellationToken ct);
 
 		void Add(T entity);
 		void Remove(T entity);
